+++ interfaces/kits/components/ui_arc_label.h
@@ -178,7 +178,7 @@ public:
      * @since 1.0
      * @version 1.0
      */
-    void SetFontId(uint8_t fontId);
+    void SetFontId(uint16_t fontId);
 
     /**
      * @brief Obtains the font ID composed of font name and size.
@@ -187,7 +187,7 @@ public:
      * @since 1.0
      * @version 1.0
      */
-    uint8_t GetFontId();
+    uint16_t GetFontId();
 
     /**
      * @brief Sets the font for this arc label.
@@ -358,7 +358,7 @@ protected:
                              TextOrientation orientation);
 
     virtual Rect GetArcTextRect(const char* text,
-                                uint8_t fontId,
+                                uint16_t fontId,
                                 uint8_t fontSize,
                                 const Point& arcCenter,
                                 int16_t letterSpace,
